+++ src/lib.rs
@@ -2,16 +2,6 @@ pub mod frontmatter;
 pub mod providers;
 pub mod siena;
 pub mod utils;
-pub mod yaml;
 
 #[cfg(test)]
 mod siena_tests;
-
-#[cfg(test)]
-mod frontmatter_tests;
-
-#[cfg(test)]
-mod utils_tests;
-
-#[cfg(test)]
-mod yaml_tests;